An accessibility improvement
I decided that colored books were a bad choice for a few reasons. The first reason is that books don't directly translate to licenses. Colors also make it harder for people with color blindness to clearly see the desired and intended answer to the question. I think making them check marks, minus signs, or Xs will make this far easier to read and make it easier to maintain as well since it is more consistent with the rest of the document. I hope this helps, a lot of people were requesting this!